# app/api/admin/quiz/[quizid]/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { getQuiz, updateQuiz, removeQuiz, assertOwnsQuiz, getEmailFromAuthorization, save } from '@/app/lib/service';

export async function GET(
  request: NextRequest,
  { params }: { params: Promise<{ quizid: string }> }
) {
  try {
    const authHeader = request.headers.get('Authorization');
    const { quizid: quizId } = await params;
    if (!authHeader) {
      return NextResponse.json({ error: 'No authorization token' }, { status: 403 });
    }
    const email = getEmailFromAuthorization(authHeader);
    await assertOwnsQuiz(email, quizId);
    const quiz = await getQuiz(quizId);
    return NextResponse.json(quiz);
  } catch (error: unknown) {
    if (error instanceof Error && error.name === 'AccessError') {
      return NextResponse.json({ error: error.message }, { status: 403 });
    }
    return NextResponse.json({ error: 'A system error occurred' }, { status: 500 });
  }
}

export async function PUT(
  request: NextRequest,
  { params }: { params: Promise<{ quizid: string }> }
) {
  try {
    const authHeader = request.headers.get('Authorization');
    const { quizid: quizId } = await params;
    if (!authHeader) {
      return NextResponse.json({ error: 'No authorization token' }, { status: 403 });
    }
    const email = getEmailFromAuthorization(authHeader);
    await assertOwnsQuiz(email, quizId);
    
    const { questions, name, thumbnail } = await request.json();
    await updateQuiz(quizId, questions, name, thumbnail);
    await save();
    return NextResponse.json({});
  } catch (error: unknown) {
    if (error instanceof Error && error.name === 'AccessError') {
      return NextResponse.json({ error: error.message }, { status: 403 });
    }
    return NextResponse.json({ error: 'A system error occurred' }, { status: 500 });
  }
}

export async function DELETE(
  request: NextRequest,
  { params }: { params: Promise<{ quizid: string }> }
) {
  try {
    const authHeader = request.headers.get('Authorization');
    const { quizid: quizId } = await params;
    if (!authHeader) {
      return NextResponse.json({ error: 'No authorization token' }, { status: 403 });
    }
    const email = getEmailFromAuthorization(authHeader);
    await assertOwnsQuiz(email, quizId);
    await removeQuiz(quizId);
    await save();
    return NextResponse.json({});
  } catch (error: unknown) {
    if (error instanceof Error && error.name === 'AccessError') {
      return NextResponse.json({ error: error.message }, { status: 403 });
    }
    return NextResponse.json({ error: 'A system error occurred' }, { status: 500 });
  }
}